requirements.txt
pytest==7.4.2
requests==2.31.0
Faker==19.6.1
allure-pytest==2.13.2
allure-python-commons==2.13.2
selenium==4.12.0